Default `convert` output to system `/tmp`
Modifies the `convert` command to output PDFs to a subdirectory
within the system's temporary directory
  `os.tmpdir()/md-to-pdf-output`
by default, when `--outdir` is not specified.
The full path to the generated PDF is logged to the console.

This change aims to reduce clutter in the user's source or working
directories.

Old workflow (implicit, without `--outdir`):
  md-to-pdf convert mydoc.md
  # PDF appears in ./mydoc.pdf or ./<source_dir>/mydoc.pdf

New workflow (implicit, without --outdir):
  md-to-pdf convert mydoc.md
  # PDF saved to temporary directory: /tmp/md-to-pdf-output/mydoc.pdf

This helps keep pdf's from flaking around the file system for basic
testing in live documents.