md-to-pdf-cm: Implement CollectionsManager module
feat(collections-manager): Add 'add' and 'list downloaded' commands
- Implements `CollectionsManager` class in
    `src/collections_manager/index.js`.
- Adds functionality to clone collections from Git URLs or copy from
  local paths.
- Collections are stored in `COLL_ROOT` (default: XDG data directory).
  - TODO: this will eventually follow same hierarchy as md-to-pdf's
    trumping rules for configs and plugins
- Implements `list downloaded` to show collections in `COLL_ROOT`.
- Creates a dedicated CLI runner `collections-manager-cli.js` using
  yargs.
- Adds `fs-extra` and `chalk` (v4.1.2) as dependencies.
- Includes basic unit tests in
    `src/collections_manager/test/run-cm-tests.js`.

Because this environment is smaller that the main md-to-pdf repo, we are
introducing stdout colors with chalk.  The goal of this module is also
not to tangle its logic on the main repo's structures. This way, this
module could effectively be generalized as a bespoke plugins manager for
node.js projects.  Meanwhile, once the core features of md-to-pdf-cm are
functioning and usable, `md-to-pdf collections` will be implemented, and
will necessarily depend on `md-to-pdf-cm`.